reddit-client/requirements.txt
praw==6.4.0
kafka-python==1.4.7